Ashley/requirements.txt
openai==2.6.1
azure-ai-projects==1.0.0
azure-identity==1.25.1
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==21.2.0
jinja2==3.1.4
